I did not make any headway on the chosen problem, but I figured out how to type the parse_n_array without having to run it once. I discovered a scheme of carrying types around that would be suitable for a language like Spiral.
The way it would work is simple. I just thought of the array_elem_type that would do as it name indicates and then generalized it to functions and type constructors.

I definitely need this kind of function if I want to implement the array library functions.

While I was working on the design for this, I also realized I could implement typec_map for mapping over the types of type constructors.

I would need this to implement the list map without having to pass in a type explicitly.

Today was quite surprisingly all in all. I remember trying to tackle this same problem before and utterly failing, but now it is coming to me with great clarity.

Type constructors are such a great feature. They will allow a young language like Spiral with just unidirectional type propagation to compete with ML big boys. They are definitely more than for just defining algebraic datatype. They are the magic ingedient that will allow types to be propagated through the program safely.

As usual at this point in the day, I am dead tired so I am going to stop here.