I am under enormous strain due to lack of rest tonight, so I will call it a day here. Right now I am thinking about the writter monad and how I might extend the parser to support it for the sake of propagating errors.
It is not really a priority though. I'll keep it for the future.

Here is a little tally of what I did today:

1) Redesigned the && and the || operators so they are short circuiting. I spent a lot of time cycling through various complicated schemes before settling on the simple way of doing it that I have now. The short circuiting also applies to typechecking, but on the other hand that should make these operators a lot more useful in pattern matching.

2) Added typec_map. With this, the type constructor mechanism essentially acts as a channel. If I added typec_bind I could use the mechanism as a monad. It might be something to consider when the time comes to write Spiral in Spiral.

3) Decided not to make `.` bind more tightly because it would mess up expressions like `ar.elem_type.elem_type`. What I have now is good enough.

4) Rewrote the parser library (for the fifth time now). Added the choice combinator.

5) Rewrote the list library. Lists are now represented as modules and have the elem_type field that acts as a type container for the generic argument.

6) Extended typec_strip so it covers modules.

7) Added eq_type as a function to the core.

My brain feels like it is on fire and not in a good way, so I will stop here, though I have a feeling that my obsession won't let me go just like that. It is kind of a double edged sword as it gives me no respite even if it would be better for me if it did.

Everything that I've done today is as of yet untested and will be left at that for the time being. There is absolutely no way I could deal with debugging right now, so I'll wisely leave it for later.

After I finish debugging this, I think these will be the last batch of big changes to the language.

With this, there will be absolutely nothing stopping me from going out and doing those HackerRank challenges as planned.

After that it is will be time for the Cuda backend and the ML library.

And after that, online poker.

If I can manage thie before the year is up, it will be quite an accomplishment.