I had a quite a good night's sleep tonight and feel quite refreshed. I've been thinking that today I should take a break from programming to read that category theory book for programmers.
https://bartoszmilewski.com/2014/10/28/category-theory-for-programmers-the-preface/

Usually things like this are almost a complete waste of time, but the part that is not wasted tends to be useful.

Also I've decided to make && and || non-short circuitable, which is what they are right now.

I am trying to balance inlining and short circuiting and am finding it an enormous pain in the ass. It is just not worth it, this level of annoyance.

...Now that I think about it, the way I've implemented delaying in && and || will make them completely non-extensible because the parser will compulsively wrap the function.

Let me instead make && and || actual keywords then...

Edit: This is quite a complicated problem, but I have it now. In the pevaller, turn the and and or into if statements and then optimize them in the codegen.

Don't virtualize them, instead treat them as keywords in standard code. Also so they can be passed around as higher order functions, do keep them as functions as well. This is the ideal way.

Also to my surprise I got a response from the HackerRank team. Let me make a little simplified example for them.