No, if I did what I had proposed then I would have no way of delaying the computation.
Fundamentally, || and && can't both be 'a -> 'a -> bool and 'a -> Lazy<'a> -> bool, but that is how the two things act in actual coding, so it is no wonder that I am getting confused here.

Of course if you think of && and || as macros then their semantics make perfect sense. So without a doubt, what I have now is the best that can be done. Let me move to testing them.

No doubt there is a painful debugging session ahead of me, but I doubt it will be that bad.

The one in the past few days was bad, but that was because I did not understand why the loop was crashing when not in tail position and that thing in sprintf really got me.

This time the extension is straightforward so I just have to test that it makes sense and then I will be done.