Made an addition to the if_body so the same branches get propagated back. Now with this, I am the point where I can't improve propagation without significant pain.
Last night I had a lot to think about how to do the list and I've concluded that my current definition of the type is absolutely flawed.

Unlike with monads, I cannot propagate the type here using elem_type.

The reason for that is that type strips all inner type constructors which would leave elem type as an instance on unboxing.

That would make it quite a different type than a plain list.

So that plan is failure. What I am going to go to is my old plan of ripping the type directly from the instance.

I tried this once before and concluded that it was impossible, but now I see quite clearly how to do it. In fact, it is quite easy.

I do not what change, probably using monads as channels made me see how type constructors are of the same kind and can be similarly manipulated. Reasoning out typec_map really made me see the light in that respect. I just need to go an extra step and make typec_slip in order to unplack those union type into individual type constructors.

Then I will just use Tuple.find to fect the relevant thing.

That having said, I've noticed that there is a need for some kind of terminating fold in the library. What I have now is ridiculous. Every time I need to exit early, I run into some kind of problem.

Well, forget that for a bit. Let me implement typec_split and then I will take care of the List module.