This is not going well at all.
I got up later than usual today, added the same branch check to the if statements, did the typec_split, spent two hours outside, and for the past two hours made only piddlying progress.

I spent most of this time getting distracted by Reddit.

Usually when that happens, it is not that I am lazy, it is just that I am completely stuck and need to rethink my approach.

Basically, as soon as I started work on the List module it became apparent to me that the way I am approaching the problem makes no sense.

The typec_map function is especially confusing since it acts like both `a -> b` and `a -> m b`.

Basically, I need to stop here and reconsider my approach.

When a function does two things at once, that is a sure sign that things are confusing. Right now, there is no problem with mapping, but I am confused how to deal with pattern matching failure.

I expected this to be...well, I did not think about it much at all. And now I realize that this is a huge issue.

What I am leaning towards is stopping work on the List mode for a while and doing a proper monadic bind for typec and then figure out how I want to deal with failure cases in typec.

Well, these kind of things happen. So much for my plans to go to dealing with HackerRank problem.

Yeah, I am not even close to being done with adding features to the language.

I need to dedicate some time to dealing with this new challenge.

I'll stop work on the List to do some more imagination programming.

As the language designer, it is my task to decide what is right and what is not. I am the lawmaker here and it is my responsibility to make good rules.

This is the only thing I have to do and the only thing I can do.