I made chain and finally did enough to finish the List module. I'll test whether it works tomorrow.
Right now I feel like calling it a day. Hard to believe I am still not done with the list, but I definitely needed to take the time to work this out.

Without a doubt, the only way out of the TypeC channel is through eq_type. I spent hours thinking about this so I am sure now.