8:15am. I must have gotten at 6am and spent the rest of the time thinking, but I had some wonderful ideas. As a result, I've figured out how to perfect renaming and speed up the pevaller by another 100x. With this I will be able to crush the polynomial times on long chains of functions like for that tuple parser test to linear time.
It comes back to the record idea that I had. Spiral sort of merged the concept of records with modules which leaves the name empty. What I will do is hold it to its literal meaning.

What records will be is renamed functions and modules. Unlike standard modules, their individual variables will not get tracked at the top level, instead they will be extracted like and reasigned on demand.

Having a data structure such as that will allow me to rename functions in the environment at join points and assign them to records. What that will allow me to do is elide renaming on nested function (because they will be inside records) and allow me to balance tracking of variables with the rest of the language features.

In particular what this will allow me to do is remove the dumb nodification I am doing right now on every single fucking data structure which causes significant performance degradation and accumulation of floating garbage. Resolving that issue will allow me to optimally balance memoization in the language and will increase the general degree of integration.

This is absolutely spectacular. It will also allow me to easily add heap records as well.

I will learn from this lesson and from here on out, never assume that the language is finished.

I won't start work on this right away as the language is currently broken and I need to finish testing List and the new Parsing modules first. But after that I will do it.

And after I do that, with the language at that optimal point, a rewrite of it in itself will actually have a chance of significantly removing the rest of the cruft.

This is it. I do not need those retarded GC schemes I've been trying to figure out. All I need is to handle what I have better. A slight algorithmic improvement can go really, really far in terms of effectiveness.