Testing from the top, I saw that test13 was failing. Upon closer examination I realized that the way let statements work in general is completely broken and block partial evaluation when pattern matching.
This is due to a decision I made in the previous version while I had that useless optimization pass.

Now it is obvious to me that I could have handled this a lot better. Fixing this won't be a problem and it will both speed up the compiler due to less nodes being created and make the Typed AST more readable.

This took me quite a while to figure out despite being so simple. I can be surpringly difficult to notice an error that stems from broken assumptions.