Only the last gives me the last expression needs to be unit nonsense.
I'll deal with it tomorrow.

Today I actually put in quite a bit of overtime.

I feel more confidence in the language now that I've gone through a testing spree. I will tackle the last test tomorrow and hopefully finally make some headway on the List module.

For some reason it is turning out to be quite the monster.