I am finally, finally done with the List module. There were so damn many bugs to go through I thought I would choke from the bug spray.
The next thing to deal with would be to test the new Parsing module.

After that I can move on to adding two new major features. Records and removing type constructors.

Records would allow me to finally crush the compile times down to linear regardless of the case and vastly improve the efficiency of the compiler.

Changing the type system so at to allow TyT to be passed around without being wrapped in a constructor or being a unit would allow me to resolve some things that have been troubling me for a while.

It would not address any of the pain points in the List module, but the way type constructors operate makes me uncomfortable, especially with regard to type equality.

The current system seems to work, but I sort of made it up on the fly. I should probably work on it more before I can really call it complete.

Yeah, it will be a while yet until I can start working on HackerRank problems without interrpution. I did not even touch the array functions.

...Hmmm...to be honest, I don't even care about Parsing anymore. Having to go through another ball busting debugging session only to tear everything down again hardly sounds fun to me right now. Yesterday was a slaughterfest.

How about this?

How about I implement naked type instances and records and then I deal with everything else?

Spiral seems to be stuck in a limbo where it is always two weeks away from completion, but this time I mean it.

Once I do this, Spiral really will have all the core features that it needs. Yesterday I've been inspired by an argument saying that languages which have longevity are the ones which do not dictate how their components are used.

I'll go that way then.