"Forget this thing. It is now worth thinking about. Maybe it is a strange GC issue, maybe something I did really did make the pevaller less efficient.
It should get better once I bring in records and remove excessive nodification."

The reason it went from 1.33 to 3.8 is because I was running it in Debug mode. I am not sure why it went to 1.4 instead of all the way back. When I first run it in Release it finished in 1.35, but now I can't get it to go below 1.4. The more I repeatedly run it, the slower it gets for some reason.

Forget it. Well, now the new thing works and I can confirm that all the tests pass.

I guess it is finally time.

Let me start work on records. That will be the last big feature of Spiral. It will take me a bit of time to get right, but it will be time well spent.

Time to get going.