Figured out what was wrong with test41. Actually it is quite remarkable how correct the compiler was so far given the magnitude of the changes.
Let me go through the rest of the tests and then I will consider what to do next. One thing I am going to do is embedd the free variable list into the type so it is unecessary to recalculate it and so the unboxer has an easier time on application.

But that are just details. The great achievement is putting in records in the first place.

If I showed the language as it is now to the me of few months ago, he would be astounded.

Just wait until I finish the ML library perfectly.