Hmmm...hmmmm...
With this, the records are in! I've fixed the error in the last test that was due to the autoconversion of the `list` function at the join point and now it all works the same as before.

With this I think I've essentially put the last two great features into the core Spiral - term types and records.

With this, the Spiral language is now complete, though imperfect.

It passes the gauntlet of tests, but it really needs many more before it has much full confidence. And besides that, I will need to remove some things like ForCastT and replace it with just the `term_cast` function.

In essence with records I've essentially gone full circle with the language. For the first two or three months while I was working on it, I essentially did not understand renaming, and the old Spiral version tracked variables through the type system.

This had all sorts of limitations like the inability to return functions from join points. When I after many months finally figured out that I did not need to track variables inside types and replaced that with the TyFun structure, that was a major accomplishment for me. Every step I go forward like that, it signifies that my understanding of language design has advanced as well.

Right now I am thinking what to do next.

There is a laundry lists of tasks left for me:

1) Improve the efficiency by removing the Func active pattern.
2) Replace the ForCastT type with a standard function term_cast.
3) Tests for Array.
4) Tests for Parsing.
5) Add the UniqueTypeIdT and redo List with that.
6) Add rebinding of type literals. In essence add the lit_cast and add a pattern for that.
7) Resume the HackerRank problems.

I think I will start by doing #3. I've yet to put the Array module through the paces and that needs to happen.

I need to dedicate some time to going through the Pharo course as well. When it starts, I think I will go through it.

The notion of everything being an object did inspire me to make one change to the language. I've modified the recordify function so it does not trigger on a function with just one variable.

Should I ever want to go the OO route, that will allow me to wrap primitives in functions without any kind of overhead. Quite nice. All the operators except from && and || are already virtualized anyway.

Spiral is extremely powerful in that it can emulate the object orientation of dynamic languages without any overhead.

...Just now I finished modifying the is_static function. Maybe I should make dynamize a mirror of it? ...Nope. This is good enough...

I think that just like I did right now, I will go at it bit by bit. The greatness of completing the language aside, the mop up will take quite a while.

The battle is not over. It is only getting started.

The language will be a great success - either it will kill C, or using the perfected DSLs at my disposal that I will make with it, I will kill my opponents in the AI game. I will not get beaten by those deadbeats still stuck on 70s technology.