Done with testing arrays. As expected there were a few bugs to fix. I decided to remove the automatic recordification at join points.
What happened in Array map, is that it would convert itself and cause all of its arguments to be passed as duplicates. Not what I had in mind.

It made me consider briefly what I was doing, but in the end, I decided that finding the correct placement for the join points was probably a too difficult to leave to the compiler anyway.

In the Parsing module, I will just have the bind function do it instead. That should work well at any rate. The same goes for term casting.

This kind of thing should be better left to the user.

It is still a far cry from C.

C brings you closer to the metal and Spiral brings you closer to the optimizer. Other languages might let you make DSLs, what Spiral will do is allow you to perfect them.

I am thinking what I should do to speed up the renamer and I think I am going to bring back memoization. What I will do this time is employ reference equality instead.

The core functions that will be the peak of the overhead in deeply nested scenarios will always be the same anyway so I should take advantage of that.

Let me do it now and I will call it a day.