I see that the previous commit did not go through. No matter, I am ready to start work on Parsing.
The truth is I really should be putting in the Writter pattern for the sake of error handling, but I need to do it like this first so I can get a feel for performance. I am going to have to redo the library too since I changed how types are propagated.

Sigh...well, just this. I am not that far away until I am 'done' with necessary, but tedious fundamental work on the language and will soon be able to make real programs.

Making a language such as Spiral is something that one experiences only once in his life, so I should savor these brief moments.

Let me continue walking the path.