As usual, debugging the Parsing module is like pulling your teeth out.
Spiral really shows its weakness here with it lack of type inference.

I think that adding Spiral's core to a ML like language would be the ideal way to combine the strengths of the two languages, but I am going to leave that for later.

Most of the parsers work, but parse_n_array is still giving me trouble. I'll deal with it tomorrow., right now I lack the energy.

It is time call it a day.

...Actually, I just figured out what is wrong with parse_n_array. Let me deal with it.

...Ok, I did but now there are codegen errors.

I'll leave it for tomorrow. That last thing was trully the final burst.