1) Tests for Parsing. 2) Resume the HackerRank problems. 3) Figure out why the Trace for TyT is not working. 4) Add the [] syntax to the language. 5) Improve the error for the pattern matching miss. Print out what the argument is. 6) Remove the `module()` and extend the stardard {} so I can put in variable names directly. 7) Change the semantics of semicolons so the blocker is line sensitive. What I have now makes no sense and is incompatible with F# anyway. It is ugly as well. 8) Replace is_static with is_lit.
I am done with the Pharo videos. I am both quite impressed with the syntax of the demonstrated DSLs, and the tooling.

http://www.cas.mcmaster.ca/~carette/publications/scp_metamonads.pdf

Right now, I am reading this paper on staged monads. After I am done with this I will resume programming.

There are a bunch of thing I need to do as can be seen from the above list. I've removed the active patterns and the unique type id thing. Forget that for now.