https://www.youtube.com/watch?v=ecIWPzGEbFc
Somehow I ended up the redesigning the parsing module while listening to this talk. It is quite preachy and trip into nostalgia rather than a vision into the future which the title promised. It is uninspired garbage.

I never program while listening to things on the side since it distrupts my concentration, but it happened today.

This will be the last redesign where I just move the arguments around.

It is just too time consuming to do it this way.  I am going to upgrade my testing process. This is clearly unsustainable.

Since I do not get direct IDE feedback I absoutely need something like that.

Probably I've wasted so much time but not doing it this way from the start, but before, I could never really visualize how to do it, since how would one test the fucking compiler?

All the tests that I've seen up to now in unit testing tutorials is testing for trivial bullshit like 2+2=4.

Rob Martin's previous talk on what killed Smalltalk is poignant to me.

But the answer is not testing. Testing is a problem - it is just too tedious. I need to lock into the differences from change to change. That will tell me whether I am on the right track.

Also the fact that it takes me hours to just move a few arguments in essentially 150 lines, while it would take me 10m in F# is an indication of just how wrong things are.

The referenced paper by Olen on abstract interpretation is surprisngly way less sophisticated than what I have now in Spiral already, though it did make me realize that I could put in some more optimizations for arithmetic operations.

Well, it does not particularly matter though, putting those things in would be easy.

So far though, the biggest problems for me were making requisite conceptual breakthroughs. This is what took me all this time for over a year.

I never felt that testing was the big overhead, so it probably was not...but doubting myself is good.

Sometimes, regrets need to be seeded an nurtured so they give a good harvest into the future.

Right now it is 9/18/217.

I got this computer in mid January of 2015 and spend 3 weeks playing games preparing my mind before launching into the journey of programming.

It is not like 2.75 years are enough for me to know everything.

Without a doubt, I've been rellying on my tooling way too much. F#'s type inferene is just so good that in different domains it might be hindering my growth.

Imagine if me doing proper testing meant that I could be where I am now 3 months ago or even more.

I will finish testing the parsing library tomorrow by hand and then I will go through the tests once more from the top to make sure that everything in set. Then I will go to the diff based approach to testing.