Everything works just so horribly. Even with term casting it is slower than it was without before. And the type propation scheme that I invented causes exponential blowups in compile times.
The more I try to make this work, the more my dreams unravel. Records are so useless that it hurts.

Just why is it so slow even with term casting?

I think I should get to the bottom of this before I make any conclussions. But records suck so much that I am going to remove them from the langauge.

I'll give it a week or two before making the decision.

https://github.com/backtracking/ocaml-hashcons

It might be worth looking up stuff on Hash Consing.

I'll continue this tomorrow when I am more rested. The first thing I will try is profiling.