Honestly, I do not feel like doing much anymore for the rest of the day, but doing all this research has helped sharpen my sense of where I should go.
https://github.com/backtracking/ocaml-hashcons/blob/master/hashcons.ml

What I should do is port this library from Ocaml to F#. Since hash consing has become such an urgent need for the language, if I want it to be successful, it is the least I can do for it.

This won't be easy. Besides rewriting the 700 LOCs to F#, I will have to figure out how weak references work and such.

But I have to do it. Right now I am getting literally the worst of all worlds, if I rewrote the thing in C and used an allocator that never freed used nodes, I would get something vastly superior to what I have now, that is before.

The current method that I am using just leaks memory and never frees anything anyway, but has the disadvantage of creating useless nodes and calculating the hash twice.

It is nonsense. There is an enormous amount of GC churn in general, in the language.

I am going to have to redo Spiral in Spiral eventually, but for now let me do what I can.

I lack power to do all that is needed and I need help, but until reinforcements arrive I will have to make do with what I have.

Today I did not do much programming and mostly did research instead.

I'll stop updating this repo for while in order to work things out.