After debugging it for 30m I've concludded that I am going to take out Patricia trees. My add hoc way of adding the c parameter like in Okasaki's paper did not work and I am not sure how to fix this.
For some reason the inter function is not working and due to that, the entire thing is worthless and just more weight on my shoulder.

Also, Jean-Filliatre said that he did not observe any benefit from using Patricia trees vs OCaml's built in AVL tree even on union operations, so that cliches it.

Let me remove it and go back to standard map.