I did what I said I would in the last commit about redoing the table from scratch, and now the table works even with weak references. When the size is set correctly, it works as well as a regular Dictionary. It is amazing how fast the thing is given that it also has to do resizing, but probably resizing is not even 0.01% of the overhead.
I will call it a day here and leave resizing for tomorrow.

It has been a struggle, but it is also starting to turn satisfying. I am learning a lot about how dictionaries and hash tables work here, in addition to hash consing.

I never thought I would find a use for this, but here I am.

I do not want to badmout Filliatre's code, probably I made a mistake somewhere in the translation, but now that I am internalizing my understanding of hash consing, I am finding his table implementation to not be lean enough. There is a bit too much duplication for my taste, and I think I can do better.

The way I just plugged that WeakArray in there is just too haphazard so it is no wonder I am getting problems, but the way his code was structured forced my hand somewhat.

Well, it is all fine and good in the end.

I'll be done with this by tomorrow at any rate.