Finally got the hash consing to work perfectly. Since the term casting speed test takes to long, I am going to try consing the rest of the structures and seeing what happens. Actually, probably just TyV should be enough.
For now, let me take a break. This was some sensitive work, I am glad to have gotten it out of the way.

With this scheme I won't have to worry about memory leaks anymore.

I looked at the CLR dictionary, but it is kind of a corporate mutable mess. It is not something I can internalize. Looking at it just reinforced my notion of how necessary proper abstractions are.

Filliatre's hash consing table implementation is quite elegant in comparison. I am glad I had the change to learn how to make a weak dictionary now.