"12:50pm. I consified TyV's type field and TyVV. Let me give it a shot now.
...It takes much longer than before.

12:55pm. Ugh, fuck.

This is going to be a pain in the ass, there is no doubt about that.

Maybe I will undo what i did just now and consify only Ty's. Maybe a problem is that it recreates the types constantly?

Honestly, this kind of refactoring is incredibly tedious and I certainly don't feel like doing it. But let me go forward with it. I need to put this out of the way."

Let me save here and I will go to the previous commit. This time I will consify each of the `Ty`s