"12:50pm. I consified TyV's type field and TyVV. Let me give it a shot now.
...It takes much longer than before.

12:55pm. Ugh, fuck.

This is going to be a pain in the ass, there is no doubt about that.

Maybe I will undo what i did just now and consify only Ty's. Maybe a problem is that it recreates the types constantly?

Honestly, this kind of refactoring is incredibly tedious and I certainly don't feel like doing it. But let me go forward with it. I need to put this out of the way.

I really wish F# had proper active patterns.

It would literally cut the time I spend refactoring this to minutes from hours.

1:30pm. I had to do chores.

While I was getting some light exercise, t got me wondering whether I remember correctly about the 0.7 thing. I looked back just now in the journal and it was more of an reduction from 4.65 to 0.7. This is a 6.5 fold increase in speed. Going from 9.8 to 2.4 seems like crap in comparison, but it is still 4x. So term casting does work. Maybe it has something to do with the program structure, who knows.

Honestly, I had enough of this.

I still have absolutely no idea how much of a speedup I could get if I cut down on allocation, but for now I had enough. I do not feel like refactoring anymore. I really got what I came for here.

If F# had proper active patterns and first class staging like Spiral, I could make a baseline, but for now I will call it a close here."

I'll merge this to master, and maybe I will profile the test once just to see what comes out, but after that I am done.

I need proper performance assurances, which means having both first class staging and first class hash consing. When I have that I will be able to maximize sharing and minimize meaningless allocations.

And the fact that I would ever have such huge chains of nested function is unusual - this scenario certainly won't happen in the Cuda backend, and probably won't with the machine learning library either.

Let me just finish debugging, make the last of the improvement for the language from that list, finish the ML library and then I will be able to release Spiral 0.1, the first public version.

If it proves its worth, I will have plenty of time to rewrite the language in itself later.

Let's finish this once and for all time.