1) Resume the HackerRank problems. 2) Figure out why the Trace for TyT is not working. 3) Add the [] syntax to the language. 4) Improve the error for the pattern matching miss. Print out what the argument is. 5) Work on print functions for the raw nodes. No more raw AST dumps. It is past the point where that is useful.
Done removing TypeInType. I'll work on 2-4 tomorrow and then once that is done move to #5.

It won't be long now until I am ready to do real programming.

I've been saying this over and over like a broken record for months now, but really, barring any great insights on how to speed up the compiler, there is only so much can do here.

It will soon be time to program.

As it is late and I've been at this since 8:30am, let me stop here. Time for play.