1) Resume the HackerRank problems. 2) Figure out why the Trace for TyT is not working. ~3) Add the [] syntax to the language. 4) Improve the error for the pattern matching miss. Print out what the argument is. 5) Work on print functions for the raw nodes. No more raw AST dumps. It is past the point where that is useful.
I added the named tuple syntax to the language. It is just some syntactic sugar, but it should serve very nicely.

The language looking nice is important to me, and will be to other users. There is simply no reason not to pretty it up.

The next thing on the list will be #2 - I need to figure out why the trace is not being propagated correctly in TyTs.

After that I will take the opportunity to redesign the pattern_compile so as to remove laziness from it and improve error messages.

When I originally put laziness in, it was while I was working out partially evaluated pattern matching for the first time and it was a completely new experience for me, not to mention a major breakthrough. It was just a little safety measure to make sure everything worked.

Now, I do not need it at all anymore.

I'll start promptly after breakfast.