Much like yesterday, I did jack in the last 4 hours due to the huge, laborious chore that needs to be doing during my best programming hour in preparation for winter.
On the bright side, at this pace, we will be done in 3 days. This pace is quite a bit more than I expected though.

Now where was I? Before I move on to the birthday cake problem, an idea of extending the [] syntax came to me. Instead of just variable names, how about I do it for any kind of literal?