Done with the first project Euler problem. There was an issue where I started suspecting hash consing, but it turns out this was me passing the old state through the join point. I always make that mistake.
At any rate, this is good. The loop function is going straight into the library. It is not a trivial thing to write a loop in Spiral as it is in languages without staging.

Edit: I started working on the `for` function more.

I'll move it into the library and write proper tests for it.