https://projecteuler.net/problem=3
Uh...this is not such an easy problem considering I have to factor a number with 12 digits. But it should be doable on my machine.

Well, it should be much easier if I just take the square root as a starting point and round it down.

But that is not the point. Now I have an essential problem that needs addressing – how do I do conversions in Spiral? I still haven't implemented this. In addition, I need to figure out how to access the F# libraries from within Spiral.

I need to deepen the .NET integration.

...You know since this will be a large undertaking, let me keep the problem for later. Let me play with Pharo instead.