Doing Euler #3 took me a while. I really put in overtime today.
I did a fairly massive amount of redesigning of the for loop and added the unsafe_convert function to the language.

In the Loops module, I added a for loops which is breakable. I am not entirely satisfied with its design, but I'll leave improving it for later.

Bit by bit, the language is shaping up into becoming better and better. Tomorrow, maybe I will try Save The Princess.