A new day is here and I am done chilling.
Before I start work on Euler #4, there is something I am wondering about - does negate really need to be where it currently is, since dyn <| -1 does not work, is there a better way of parsing those minuses.

This is such an unbelievable pain in the fucking ass. Parsing unary operators is a fucking curse.
        let rec expressions expr s =
            let unary_ops =
                [case_lit_lift]
                |> List.map (fun x -> x (expressions expr) |> attempt)
                |> choice
I forgot what the reason was, but is there any point in not having negate along with lit_lift?

Let me deal with this thing. I want to get the unary operators out of the way. They are annoying beyond all reason.