Git is just being such a piece of shit. Let me just move to this branch.
What I was saying in the commit that got eaten, is that as soon as I replaced the fold functions with the loop that I built, everything broke down in ways that I did not anticipate.

Forget the HackerRank problem. I am going to clear my head and just focus on cleaning up a bit. Thankfully there is not much to redesign in array, but I need to do it clear-headedly.

When I am done I'll merge this with master and deal with it that way.