Actually, until just recently, it never even occured to me to use the IDE to compare the diffs. How foolish I am.
I like this.

Yeah, now I have a good flow. I will get this done and then move to Save The Princess.

As a matter of convetion, I have decided that outer funtcions will not be responsible for dynamizing the state. The same goes for the while loop.

I will start there.