Yesterday when I decided to call it a day, I started reading Knight Run and became completely entraced by it. The drama of the first 40 chapters was putting me to sleep, but then I could not put the thing down.
It is very uncharacteristic of me to read something until midninght and then pick it up again the next until I am done.

Usually I would spend at least some time doing programming related stuff in the morning, but today I did nothing but read until I caught up.

It is 1:30pm now. Holy shit, the second arc of the manga was great.

...

But that is that. It is time to start again. Let me first figure out how to restore master. This situation I found myself in is beyond annoying.